requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
lxml==4.9.3
